+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Unreleased
 
 ## 8/1/2021
+- Add MySQL example.
+- Add MySQL and MongoDB drivers for connection.
+- Add spooked, logs, dependencies and database managers.
 - Start refactor project.